PASN: Calculate Auth1 hash after PTK derivation
It is more convenient to wait with the Auth1 hash calculation to avoid
having to figure out which hash algorithm to use before the PTK is
derived. Auth1 hash is defined to use the same hash algorithm as the one
that was used during PTK derivation. This requires a bit more memory,
but that is justifiable with the simplified implementation.

Signed-off-by: Jouni Malinen <jouni.malinen@oss.qualcomm.com>
CRs-Fixed: 4390596
Git-repo: git://w1.fi/hostap.git
Git-commit: 5e5aba3dc8c5b3d8dd797c3b7d4d019b89b773f0
Change-Id: I282110090fedabd6bd82e65ee7e718536b1f887d